+++ packages/orq-rc/src/models/operations/getagent.ts
@@ -946,7 +946,7 @@ export const GetAgentKnowledgeBases$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01K5TY59C4Q34DSFTF78SH24G6"),
+  id: z.string().default("01K5V5RXTS9DNDHJ4WCEAP1S3X"),
   knowledge_id: z.string(),
   configuration: z.union([
     z.lazy(() => GetAgentConfiguration2$inboundSchema),
@@ -973,7 +973,7 @@ export const GetAgentKnowledgeBases$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   GetAgentKnowledgeBases
 > = z.object({
-  id: z.string().default("01K5TY59C4Q34DSFTF78SH24G6"),
+  id: z.string().default("01K5V5RXTS9DNDHJ4WCEAP1S3X"),
   knowledgeId: z.string(),
   configuration: z.union([
     z.lazy(() => GetAgentConfiguration2$outboundSchema),