+++ packages/orq-rc/RELEASES.md
@@ -3819,3 +3819,13 @@ Based on:
 - [typescript v3.13.0-rc.28] packages/orq-rc
 ### Releases
 - [NPM v3.13.0-rc.28] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.28 - packages/orq-rc
+
+## 2025-09-23 11:38:06
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.622.1 (2.709.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.13.0-rc.29] packages/orq-rc
+### Releases
+- [NPM v3.13.0-rc.29] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.29 - packages/orq-rc